+++ KursowWork/src/com/mashka/views/Enter.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane maxHeight="260.0" maxWidth="300.0" minHeight="260.0" minWidth="300.0" prefHeight="260.0" prefWidth="300.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.EnterController">
+<AnchorPane maxHeight="260.0" maxWidth="300.0" minHeight="260.0" minWidth="300.0" prefHeight="260.0" prefWidth="300.0" xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.mashka.conrollers.EnterController">
    <children>
       <TextField fx:id="FirstText" layoutX="46.0" layoutY="68.0" AnchorPane.bottomAnchor="170.0" AnchorPane.leftAnchor="40.0" AnchorPane.rightAnchor="40.0" AnchorPane.topAnchor="60.0" />
       <TextField fx:id="SecondText" layoutX="46.0" layoutY="130.0" AnchorPane.bottomAnchor="100.0" AnchorPane.leftAnchor="40.0" AnchorPane.rightAnchor="40.0" AnchorPane.topAnchor="130.0" />
